moving static fixtures into data_fixtures app
