+++ data_fixtures/management/commands/update_offered_by.py
@@ -2,7 +2,7 @@
 
 from django.core.management import BaseCommand
 
-from learning_resources.utils import upsert_offered_by_data
+from data_fixtures.utils import upsert_offered_by_data
 from main.utils import clear_search_cache, now_in_utc
